refactoring and reordering dataframe
